# Render unpaved roads different