… roads.mss
… symbols/unpaved/unpaved.svg
… symbols/unpaved/unpaved_aeroway-fill.svg
… symbols/unpaved/unpaved_living-street-fill.svg
… symbols/unpaved/unpaved_motorway-fill.svg
… symbols/unpaved/unpaved_motorway-low-zoom.svg
… symbols/unpaved/unpaved_pedestrian-fill.svg
… symbols/unpaved/unpaved_platform-fill.svg
… symbols/unpaved/unpaved_primary-fill.svg
… symbols/unpaved/unpaved_primary-low-zoom.svg
… symbols/unpaved/unpaved_raceway-fill.svg
… symbols/unpaved/unpaved_residential-fill.svg
… symbols/unpaved/unpaved_road-fill.svg
… symbols/unpaved/unpaved_secondary-fill.svg
… symbols/unpaved/unpaved_trunk-fill.svg
… symbols/unpaved/unpaved_trunk-low-zoom.svg