Fixes #110 using the idea from #2640 while trying to avoid re-introducing #3280