@sommerluk 2018-08-22
@rrzefox Could you test this branch on your server giving some feedback about performance on different zoom levels, compared to the current master?