@rrzefox 2018-08-23
I have tried to simply render a 2048x1024 tile with nik4.py on zoomlevels 8 to 12. The server has a full planet DB. The results are not good. Each test was run 3 times, once for warmup, and the following table gives the best of the other 2 results:
```
Zoomlevel     without_this_PR     with_this_PR
   08              82                 119
   09              30                  43
   10              53                 188
   11              25                  70
   12              15                  35
```
Under these circumstances, I really do not want to deploy this into the live rendering ;)
<details><summary>benchmark output</summary>

unpatched run 2:
URL 0 (https://www.openstreetmap.org/#map=12/49.6100/11.0262):

real    0m20.997s
user    0m5.460s
sys     0m0.528s
URL 1 (https://www.openstreetmap.org/#map=11/49.5096/11.0625):

real    0m34.405s
user    0m8.904s
sys     0m0.560s
URL 2 (https://www.openstreetmap.org/#map=10/49.5626/10.9850):

real    0m54.297s
user    0m19.280s
sys     0m0.940s
URL 3 (https://www.openstreetmap.org/#map=9/49.7049/10.8820):

real    0m29.649s
user    0m16.268s
sys     0m1.108s
URL 4 (https://www.openstreetmap.org/#map=8/49.539/10.305):

real    1m21.594s
user    0m33.612s
sys     0m1.892s
Rendering benchmark took 221 seconds total.

unpatched run 3:
URL 0 (https://www.openstreetmap.org/#map=12/49.6100/11.0262):

real    0m15.126s
user    0m5.436s
sys     0m0.544s
URL 1 (https://www.openstreetmap.org/#map=11/49.5096/11.0625):

real    0m24.917s
user    0m9.248s
sys     0m0.620s
URL 2 (https://www.openstreetmap.org/#map=10/49.5626/10.9850):

real    0m53.036s
user    0m19.380s
sys     0m1.200s
URL 3 (https://www.openstreetmap.org/#map=9/49.7049/10.8820):

real    0m31.206s
user    0m17.304s
sys     0m1.208s
URL 4 (https://www.openstreetmap.org/#map=8/49.539/10.305):

real    1m23.309s
user    0m34.644s
sys     0m2.412s
Rendering benchmark took 208 seconds total.

patched run 2:
URL 0 (https://www.openstreetmap.org/#map=12/49.6100/11.0262):

real    0m35.593s
user    0m9.564s
sys     0m1.212s
URL 1 (https://www.openstreetmap.org/#map=11/49.5096/11.0625):

real    1m11.907s
user    0m17.284s
sys     0m2.156s
URL 2 (https://www.openstreetmap.org/#map=10/49.5626/10.9850):

real    3m11.256s
user    0m40.440s
sys     0m5.844s
URL 3 (https://www.openstreetmap.org/#map=9/49.7049/10.8820):

real    0m43.480s
user    0m19.144s
sys     0m1.876s
URL 4 (https://www.openstreetmap.org/#map=8/49.539/10.305):

real    2m1.607s
user    0m42.376s
sys     0m3.752s
Rendering benchmark took 464 seconds total.

patched run 3:
URL 0 (https://www.openstreetmap.org/#map=12/49.6100/11.0262):

real    0m35.349s
user    0m9.480s
sys     0m1.204s
URL 1 (https://www.openstreetmap.org/#map=11/49.5096/11.0625):

real    1m10.434s
user    0m17.092s
sys     0m2.156s
URL 2 (https://www.openstreetmap.org/#map=10/49.5626/10.9850):

real    3m7.632s
user    0m40.872s
sys     0m5.304s
URL 3 (https://www.openstreetmap.org/#map=9/49.7049/10.8820):

real    0m43.129s
user    0m19.108s
sys     0m1.432s
URL 4 (https://www.openstreetmap.org/#map=8/49.539/10.305):

real    1m59.042s
user    0m41.592s
sys     0m3.652s
Rendering benchmark took 456 seconds total.

</details>