@imagico 2018-08-23
Based on a few quick tests without a global database it seems that even without the global bounding box geometries and the comp-op stuff the restructured MSS code is significantly slower than before.  The SQL code seems to be fine now OTOH.

It has been some time since i looked over this so forgive me if this is stupid but can't you sort the global bounding box (in several instances, one for each road color) into the SQL defined drawing order and this way get rid of the attachments in the MSS code (and this way keep the MSS structure essentially as it is now - just adding the comp-op stuff)?