@timautin 2018-08-24
@rrzefox Thanks for sharing these numbers. Could you add some information about your hardware please?