@sommerluk 2018-08-25
@rrzefox Thanks a lot for testing this!

> even without the global bounding box geometries and the comp-op stuff the restructured MSS code is significantly slower than before

That’s what I suspected earlier (see the last point of my [previous comment](https://github.com/gravitystorm/openstreetmap-carto/issues/3280#issuecomment-403271784)).

> […] can't you sort the global bounding box […] into the SQL defined drawing order […]?

I fear that I can’t. The reason is the behaviour of `comp-op`. The following occured when rendering just by SQL order without attachments:

0. We have yet various things like landcovers rendered. Let’s call this canvas A.
1. Render white unpaved roads: Cut a hole in the existing image by using comp-op. Now we have canvas A with some holes.
2. Render the white pattern behind: It will be only visible where we have the holes in canvas A that have been done in the previous step. Now we have still canvas A with some holes. And we have canvas B with the white pattern on the global bounding box. Canvas B is behind canvas A.
3. Render white paved roads. They will be rendered in canvas A.
4. Render yellow unpaved roads: Cut a hole in the existing image by using comp-op.  This cuts more holes in canvas A.
5. Render the yellow pattern behind: This will create canvas C, which will be behind canvas B. It will therefore never be visible.