@imagico 2018-08-25
That is unfortunate.

But what you describe to me kind of contradicts the mapnik docs because what you do is what in

https://github.com/mapnik/mapnik/wiki/Compositing

is described as _Feature level compositing_ which should work on a per feature basis and not involve the creation of temporary canvas (which is described below as _Style level compositing_).

The lack of reliable documentation of how mapnik works really hampers developing a solution to this kind of problem.