@kocio-pl 2018-08-25
Well, maybe you could just ask Mapnik developers then - @talaj, would you have any advise here?