@imagico 2018-08-27
Adding `cache-features: on` to all layers involved as per https://github.com/mapnik/mapnik/issues/3977#issuecomment-416269110 seems to improve performance but i have not tested this in more detail.

That's a heck of an unexpected default setting.