@kocio-pl 2018-08-28
It looks like upstream (Mapnik) implementation of roads with pattern rendering is at least doable - https://github.com/mapnik/mapnik/issues/3977#issuecomment-416357333. That would be really great...