@kocio-pl 2018-08-31
Even better - it's a real work in progress! See https://github.com/mapnik/mapnik/pull/3980.