@sommerluk 2018-09-02
Rebased.

I’ve implemented the performance tuning that @talaj had [proposed](https://github.com/mapnik/mapnik/issues/3977#issuecomment-416269110).

> Adding cache-features: on
That didn’t work. I eventually used `cache-features: true` instead.

Now, on my local machine (based in single tile rendering once per zoom level), the branch `unpaved20` is even slightly faster the current `master` branch. @rrzefox Could you confirm this?