@imagico 2018-09-02
> That didn’t work. I eventually used cache-features: true instead.

Strange - because for me it did work.

It would also make sense by the way to check if other layers would profit from this as well.  Roads is the most obvious candidate but it is not the only place attachments are used.