@sommerluk 2018-09-02
> check if other layers would profit from this as well

Indeed I was thinking about that, but I wanted to wait for the results of @rrzefox before. In general, I suppose that this option is without any negative consequences on layers that do not use attachments. If so, I could imagine add this in general to all SQL layers, regardless of their current usage. This way, we do not need to worry about. (I’m not sure about non-SQL layers however…)